third_party/blink/web_tests/platform/android/external/wpt/pointerevents/pointerevent_touch-adjustment_click_target-expected.txt

This is a testharness.js-based test.
[FAIL] Touch-generated events should have the same target
  assert_equals: pointerdown should have un-adjusted y coordinates. expected 53 but got 52.999996185302734
Harness: the test ran to completion.
